feat(dynamodb): enhance MultiResponse type to support conditional item return and bump version to 1.0.43
This commit updates the `MultiResponse` type in the `Dynamodb` namespace to include a new generic parameter, `ReturnAsPersistedItem`, allowing users to specify whether the returned items should be of type `PersistedItem<T>[]` or `T[]`. This change enhances the flexibility of the response structure, accommodating different use cases in data handling. Additionally, the version in `package.json` has been incremented to 1.0.43 to reflect this update.